change router structure
