+++ src/app/(authenticated)/users/[userId]/SettingsLink.js
@@ -10,7 +10,7 @@ const SettingsLink = ({ userId }) => {
       <Link
         className="w-full text-center font-semibold text-sm p-2 bg-primary rounded-md text-primary
   border border-slate-400 hover:border-slate-600 hover:bg-highlight sm:w-1/3"
-        href={`/${userId}/settings`}
+        href={`/users/${userId}/settings`}
       >
         Edit profile
       </Link>
