+++ src/app/(authenticated)/users/[userId]/page.js
