+++ src/firebase/firestore/Quote/addQuote.js
@@ -5,6 +5,7 @@ export default async function addQuote(quoteId, data) {
   try {
     await setDoc(doc(firestore, "quotes", quoteId), {
       createdAt: Timestamp.fromDate(new Date()),
+      id: quoteId,
       likes: 0,
       ...data,
     });
